+++ mobile/android/android-components/components/feature/addons/src/test/java/AddonManagerTest.kt
@@ -24,6 +24,7 @@ import mozilla.components.concept.engine.webextension.DisabledFlags.Companion.AP
 import mozilla.components.concept.engine.webextension.DisabledFlags.Companion.APP_VERSION
 import mozilla.components.concept.engine.webextension.DisabledFlags.Companion.BLOCKLIST
 import mozilla.components.concept.engine.webextension.DisabledFlags.Companion.SIGNATURE
+import mozilla.components.concept.engine.webextension.DisabledFlags.Companion.SOFT_BLOCKLIST
 import mozilla.components.concept.engine.webextension.DisabledFlags.Companion.USER
 import mozilla.components.concept.engine.webextension.EnableSource
 import mozilla.components.concept.engine.webextension.InstallationMethod
@@ -1003,6 +1004,9 @@ class AddonManagerTest {
 
         whenever(metadata.disabledFlags).thenReturn(DisabledFlags.select(APP_VERSION))
         assertEquals(Addon.DisabledReason.INCOMPATIBLE, extension.getDisabledReason())
+
+        whenever(metadata.disabledFlags).thenReturn(DisabledFlags.select(SOFT_BLOCKLIST))
+        assertEquals(Addon.DisabledReason.SOFT_BLOCKED, extension.getDisabledReason())
 
         whenever(metadata.disabledFlags).thenReturn(DisabledFlags.select(0))
         assertNull(extension.getDisabledReason())